pyproject.toml
[project]
name = "hemppa"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "matrix-nio>=0.25.2",
    "pyyaml>=6.0.2",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
dynamic = [
    "version"
]

[project.scripts]
hemppa = "hemppa.__main__:main"

[poetry.optional-dependencies]

[tool]
[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry]
version = "0.0.0"
packages = [
    {include = "hemppa", from = "."},
]

[project.optional-dependencies]
mastodon = [
    "mastodon.py~=1.8.1",
]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
